# server/src/config/api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LAPP - OpenAPI 3.0",
    "description": "Idk wtf im doing, but this is vibing",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "lapp.fly.dev",
      "variables": {},
      "description": "Yes"
    }
  ],
  "paths": {
...
    "/api/view/{file_path}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Viewing or downloading the file",
        "description": "Viewing or downloading the file",
        "operationId": "ViewOrDownloadFile",
        "parameters": [
          {
            "name": "file_path",
            "in": "path",
            "description": "path of the file",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A PDF file",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  }
}